+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.1.0 - 2020-04-08
+### Changed
+- Updated the php sass library from leafo/scssphp to scssphp/scssphp
+- Updated SCSS behavior to compile ALL code between {% scss %} tags to a single script
+
 ## 1.0.0 - 2018-05-05
 ### Added
 - Initial release